+++ static_report/src/components/shared/Columns/ColumnCards/ColumnStatsCard.tsx
