+++ static_report/src/components/shared/Columns/ColumnCards/ColumnTypeDetail/ColumnTypeDetailNumeric.tsx
