+++ static_report/src/components/shared/Widgets/CRColumnSummaryListWidget.tsx
@@ -8,8 +8,8 @@ import {
 } from '../../../types';
 import { getDataChart } from '../../../utils/charts';
 import { transformAsNestedBaseTargetRecord } from '../../../utils/transformers';
-import { ColumnCardDataVisualContainer } from '../Columns/ColumnCard/ColumnCardDataVisualContainer';
-import { ColumnStatsCard } from '../Columns/ColumnCard/ColumnStatsCard';
+import { ChartContainer } from '../Charts/ChartContainer';
+import { ColumnStatsCard } from '../Columns/ColumnCards/ColumnStatsCard';
 
 interface CRProfilingDetailsProps extends Selectable {
   baseTable?: SaferTableSchema;
@@ -49,12 +49,12 @@ export function CRColumnSummaryListWidget({
               />
 
               <Flex my={4} alignItems={'center'}>
-                <ColumnCardDataVisualContainer height={350}>
+                <ChartContainer height={350}>
                   {getDataChart(base)}
-                </ColumnCardDataVisualContainer>
-                <ColumnCardDataVisualContainer height={350}>
+                </ChartContainer>
+                <ChartContainer height={350}>
                   {getDataChart(target, base)}
-                </ColumnCardDataVisualContainer>
+                </ChartContainer>
               </Flex>
             </Grid>
           </Flex>
